Move `PatcherHelper.patch(cls)` from `patcher-gui.py` to `patcher.py`.
Until now, the legacy module (`patcher-gui`) had to be imported in
`mkdd_patcher.py` via `importlib`, as the dash (`-`) is not a valid
character in Python identifiers.

However, cx_Freeze would not correctly recognize this module as being
a dependency, failing to bundle/rename the module, and thus making
`mkdd_patcher.exe` fail at runtime.