Show warning when unsupported code patches are encountered in a custom track.
Currently, there is no *built-in* support for code patches, meaning that
the warning will be shown for any custom track that uses any code patch
at all.

Example of the warning when Waluigi Stadium Boundless (a modified
version that defines the **code_patches** field in the `trackinfo.ini`
file) is loaded:

![MKDD Patcher - Code patches warning](https://github.com/RenolY2/mkdd-track-patcher/assets/1853278/16dcd3cc-ea4b-4943-9994-a6665c5db7da)

Support for built-in code patches may be introduced in the future; for
now, users still need to apply binary-diff-based code patches as
separate mods, as stressed in the warning message.