Source supported code patches from code mods.
Code mods (i.e. mods with the `codeinfo.ini` file) can now list the
code patches they implement in the `code_patches` field in the INI file.

This allows the Patcher to determine which code patches are available,
avoiding to show the code-patch-needed warning when custom tracks need
any of the supported code patches.