Revise how minimap transforms are neutralized for courses assigned to unintended slots.
When a custom course originally designed to replace a certain slot is
manually edited to exchange the slot it replaces, the Patcher
manipulates certain slot-specific minimap transforms to prevent the game
from distorting the custom course's minimap.

Previously, the minimap transforms were only subtly tuned; now they are
totally removed, which should provide good results in all cases, and
opens the door for supporting the debug build in the Patcher.