Support for folder mode with custom mods.
Previously, only custom courses could be used in folder mode (i.e. with
the path to the custom course pointing to an *unzipped* directory).

The `get_file_changes()` function that is used to traverse the files in
`files/` (and, more recently, also in `audio_waves/`) was not handling
mods in folder mode properly.